Removed BodyAwareHttpController with HttpController<T> I don't know why it didn't cross my mind to begin with
