+++ Fuyu.Backend.Core/Controllers/AccountRegisterGameController.cs
@@ -6,7 +6,7 @@ using Fuyu.Backend.Core.Services;
 
 namespace Fuyu.Backend.Core.Controllers
 {
-    public class AccountRegisterGameController : BodyAwareHttpController<AccountRegisterGameRequest>
+    public class AccountRegisterGameController : HttpController<AccountRegisterGameRequest>
 	{
         public AccountRegisterGameController() : base("/account/register/game")
         {
