+++ Fuyu.Common/Networking/HttpController.cs
@@ -1,3 +1,5 @@
+using System.Threading.Tasks;
+
 namespace Fuyu.Common.Networking
 {
     public abstract class HttpController : WebController<HttpContext>
@@ -6,4 +8,27 @@ namespace Fuyu.Common.Networking
         {
         }
     }
+
+	public abstract class HttpController<TRequest> : HttpController where TRequest : class
+	{
+		public HttpController(string path) : base(path)
+		{
+		}
+
+		public override async Task RunAsync(HttpContext context)
+		{
+			TRequest body = null;
+
+			// NOTE: I'm not sure exactly how we should handle this. I imagine we still want the endpoint to be called?
+			// -- nexus4880, 2024-10-07
+			if (context.HasBody())
+			{
+				body = await context.GetJsonAsync<TRequest>();
+			}
+
+			await RunAsync(context, body);
+		}
+
+		public abstract Task RunAsync(HttpContext context, TRequest body);
+	}
 }
